Quit Prompt + Screen Tear Fix + Restore Default Firewood State
