+++ Assets/Scripts/Puzzles/Editor/FirewoodEditor.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5b533f6406c05b3479f5e5d63989ec42
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
