+++ Assets/Scripts/UI/PauseMenu/ConfirmationMessage.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c136053efc33f7b4eb8e8db3a8db39ca
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
