+++ Assets/Scripts/UI/PauseMenu/ConfirmationMessage/ConfirmButton.controller.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 07e76a46fa112104f8b248df2c2581dc
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 9100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
