+++ Assets/Scripts/UI/PauseMenu/ConfirmationMessage/PauseMenuQuitMessage.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 642483a41a32d614e8552d4525751fa1
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
